move spell tests to its own file
